Add proxy_convenience_aliases.h and add release header testing.
This was causing breakages for missing CDecl when arrays were introduced, but only for the compacted inline release header (vs. using through Bazel).

Also, add a build_test so this doesn't happen in the future. :jni_test now conditionally allows the usage of the monolithic header.

Lastly, rev the release header.

Test:'bazel test  --cxxopt='\''-std=c++17'\'' --repo_env=CC=clang' ...'
PiperOrigin-RevId: 453236675
Change-Id: I2871a3c03c78f54701300727b679838f495fc44b