Adding JniType generalisation.
All types will be unified into this single representation.  This will help improve some of the class definitions.

In a follow up CL, I will move existing types to drop their template parameters and use this instead.

I will also add an additional parameter to represent the "native context" of the variable (e.g. is it a local object).

PiperOrigin-RevId: 466371629
Change-Id: I26947b90fbdafcb96568fcbe44088612a292b431