Implementing Signature generation for new Id Type.
This implementation is only for constructors and overloads, although some preliminary work is done for fields.

It has been superficially inserted into MethodSeleciton, but in the future, eventually |FieldSelection| and |MethodSelection| will be completely supplanted by |Id| (thus consolidating signature generation).

Some of the Id logic will be split out in a future CL (but it's probably non-trivial).

PiperOrigin-RevId: 474050645
Change-Id: Iaca91c6d0d42bcadc59d64ce5bafdfdc6f8f1947