Add Conjunction metafunction.
Values are applied for each element, and are iterated over via next.

Currently, there's no support for 0 size sets, and there's no generic "begin" or "end" akin to what exists in STL. Eventually, this ought to exist, but it would be nice to have another algorithm using iterators to justify it.

PiperOrigin-RevId: 488709142
Change-Id: Ic0e36392ee6802eb9d759b1c03e2104f5c7e1246