Move string to shared system class definitions and add bytearray ctor.
Frustratingly, jstring is special, because it is both fully described in terms of a jobject, but has its own type (unlike any other type in JNI), as well as behaviour.

This forms a dependency cycle in the build graph, so, I will need to rework proxy definitons in a follow up CL to support follow up types, and jstring will strangely be the first.

For now, this enables building LocalString with a LocalObject (which is janky).

PiperOrigin-RevId: 493050785
Change-Id: If6b559617110a74c4ec14ac32d39e142ddd98f90