Deleted jobject constructor for GlobalObject.
This was supposed to be constrained in an earlier CL (using Promote and Warn), but it only offered additional constructors, it didn't actually stop incorrect construction.

This change is potentially breaking, however, it is likely only breaking in malformed programs which would crash immediately (and is thus likely to be safe).

PiperOrigin-RevId: 499323472