Rename ArrayTest to ArrayTestMethodRank1.
To prevent gaps in coverage (like jstring) I would like to also add for returns and params for both methods and fields (and their static counterparts). This file is already large, so I think it makes sense to have some separate tests.

Also drive by alphabetization of BUILD file (and removal of references to rJNI... it's all rJNI).

PiperOrigin-RevId: 509724257