`CreateCopy` is now the default for locals.
This follows from the discussion on https://github.com/google/jni-bind/issues/207.

Values passed through JNI are seemingly owned by the calling context and should not be deleted. This changes behaviour in existing applications, but at most should harmlessly leak a single local reference to a jobject.

This *could* be a break a breaking change if locally created JNI objects were being wrapped by a `LocalObject` in an inner loop, but can be obviated by using `AdoptLocal` anywhere this is done.

This also could break if you are passing nulls to objects (which is not allowed but would passively fail).

Because `LocalObject<kClass>{input_obj}` is a "default path", this is going to become the default. All existing usages can add `AdoptLocal{}` to duplicate existing behaviour.

PiperOrigin-RevId: 582053094