Add multi-dimensional field tests, `ArrayView` lifetime issue.
Previously, `LocalArray` would delete itself as it fell out of scope (as designed), but this caused issues for object arrays.

Object arrays are not contiguous so they need to have a living underlying `jarray`.

This effectively broke iteration on object arrays in fields using the following syntax, because the output of `Get` would be deleted before the loops completion.

`for(LocalArray<jobject, 1, kClass> arr : obj["rank2Field"].Get().Pin()`

PiperOrigin-RevId: 585703473