JvmRef now supports a "fallback loader" for when `FindClass` fails.
On Android, secondary threads do not have access to the application loader and there is no way to recover it.

Callers can now prime a loader (i.e. whatever loader is used by the `Activity` or `Application`). This loader can be used as a fallback when looking for classes.

This CL should resolve https://github.com/google/jni-bind/issues/190.

PiperOrigin-RevId: 596647831