Add initial pass of C++20 invocable map with `Call` operator.
This flavour no longer requires a clang extension and so should unlock other compilers.

I'm keeping this in a separate CL since it's the first time both C++17 and C++20 versions for code in use and would like to flush out build failures.

In a downstream CL I will add this to `ObjectRef` to enable the new syntax and eventually update existing usage.

https://github.com/google/jni-bind/issues/42

PiperOrigin-RevId: 712596590