Remove the requirement to call SetFallbackClassLoader on Android.
Android, in all its glory, has long had pernicious issues with secondary threads initiating calls on secondary threads without creating a "fallback loader".

This is annoying because it's not intuitive you need to call this, and simply looks like a crash. This is not in the spirit of JNI Bind.

This issue is now resolved by manually scanning for the current activity and scraping the classloader. Because this is done during `JvmRef` init, there is no need to prime the fallback.

PiperOrigin-RevId: 849804699